Schedule/requirements.txt
curl==2017.0.1
Django==1.10.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
httpie==0.9.9
Pygments==2.2.0
PyJWT==1.4.2
requests==2.13.0
wheel==0.24.0
